Move tests_blanks into package blanks
