+++ openprocurement/auctions/flash/tests/blanks/auction_blanks.py
