+++ openprocurement/auctions/flash/tests/blanks/complaint_blanks.py
@@ -31,7 +31,7 @@ def create_auction_complaint_invalid(self):
     self.assertEqual(response.content_type, 'application/json')
     self.assertEqual(response.json['status'], 'error')
     self.assertEqual(response.json['errors'], [
-        {u'description': u'Expecting value: line 1 column 1 (char 0)',
+        {u'description': u'No JSON object could be decoded',
             u'location': u'body', u'name': u'data'}
     ])
