+++ openprocurement/auctions/flash/tests/contract.py
@@ -6,7 +6,7 @@ from openprocurement.auctions.core.tests.base import snitch
 from openprocurement.auctions.flash.tests.base import (
     BaseAuctionWebTest, test_auction_data, test_bids, test_lots, test_organization
 )
-from openprocurement.auctions.flash.tests.contract_blanks import (
+from openprocurement.auctions.flash.tests.blanks.contract_blanks import (
     # AuctionContractResourceTest
     create_auction_contract_invalid,
     create_auction_contract,
