+++ openprocurement/auctions/flash/tests/migration.py
@@ -5,7 +5,7 @@ from openprocurement.auctions.core.tests.base import snitch
 
 from openprocurement.auctions.flash.migration import migrate_data
 from openprocurement.auctions.flash.tests.base import BaseWebTest
-from openprocurement.auctions.flash.tests.migration_blanks import migrate, migrate_from0to1
+from openprocurement.auctions.flash.tests.blanks.migration_blanks import migrate, migrate_from0to1
 
 
 class MigrateTest(BaseWebTest):
